requirements.txt
pillow>=10
pycryptodome>=3.20
